Remove old tests and add check for the nomination API
